Add useApp customHook
